requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
websockets>=12.0
numpy>=1.26
pandas>=2.1
scikit-learn>=1.4
torch>=2.2
openai-whisper>=20231117
shap>=0.44
plotly>=5.19
ortools>=9.9
networkx>=3.2
elasticsearch>=8.12
influxdb-client>=1.41
psycopg2-binary>=2.9
pyodbc>=5.0
azure-cognitiveservices-speech>=1.36
textblob>=0.17
langgraph>=0.1.7
python-dateutil>=2.8
rich>=13.7